Change of the projects structure from monolithic to modular, and adding classes. Moved configuration to a separate yaml file. Changed environment file extension (from yml to yaml)
