+++ .gitignore
@@ -1,6 +1,7 @@
 /processed_arrays/
 /out_memmap/
 /filtered/
+/__pycache__/
 polygons_processed.*
 *.log
 .DS_Store
